Add optional HTTPClient construtor param, Add Hostname variable
Add option to pass a HttpClient to the Nanoleaf client as a constructor, in case a user (like myself) has a shared HttpClient in their application they'd prefer to use in lieu of a dedicated one.

Remove extra Nanoleaf client constructor, create one overloaded version with default values.

Add a "HostName" getter for our nanoleaf client.